// src/app/core/services/model/initialProductFamilyRoutePrioritizationRequest.ts
/**
 * Mi API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface InitialProductFamilyRoutePrioritizationRequest { 
    familyId?: number;
    routeId?: number;
}